vcs: Add `__GIT_COMMITDATE_UNIX__` to git.sh
This exposed variable contains the commit date and time as UNIX time
stamp. Since this is the number of seconds since 1970 in UTC it is a
time stamp with the following properties:

- monotonically increasing in the Git history (not true for Git "author
  dates")
- there is exactly one value per commit (hash)
- time zones do not play a role since UNIX time stamps are always UTC
- easily comparable by software because it is just an integer
- almost not human readable

This is an addition to the already existing `__GIT_COMMITDATE__` which
contains the year, month and day in a human-friendly format.

Signed-off-by: Jan Krautmacher <jan@krautmacher.org>
Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>