Fix: vcs-integration: Use UTC for __GIT_COMMITDATE__
The previous version used the local time of the machine which runs the
build. This breaks reproducible builds which is discouraged.

Likely this bug was never detected because the output format does not
contain the time but only the date. Nevertheless it is still a bug and
could make the build on certain machines, in certain time zones and
certain times not reproducible which is really hard to catch.

Signed-off-by: Jan Krautmacher <jan@krautmacher.org>