vcs-integration: Add `__GIT_HASH_FULL__` to git.sh
The git.sh library so far only supported a fraction of the hash as
`__GIT_HASH__`. This is not changed to maintain compatibility. The
`__GIT_HASH_FULL__` is provided by this commit for the case that the
full Git hash is required.

To avoid code duplication the "private" function `_git_hash_` was
replaced and not complemented by the new `_git_hash_full_` function.
Nevertheless both variables `__GIT_HASH_FULL__` and `__GIT_HASH__` are
exposed.

Since those private functions should not be used anyway this should not
affect other projects / compatibility.

Signed-off-by: Jan Krautmacher <jan@krautmacher.org>