vcs: Add generate-metadata-yaml
This script is similar to generate-version-h but is better suited to use
the shared git.sh library to instantiate YAML templates with metadata
about the code base.

Maintainer's note: We are taking this as is for now, but the generation program
should really reuse the same facilities the generate-version-h program esta-
blished; not only to reduce code duplication, but much more to have feature pa-
rity between the two generators.

Signed-off-by: Jan Krautmacher <jan@krautmacher.org>
Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>