vcs: Remove dependency upon 'date', fix possible inconsistency
The way this worked was to ask git for a unix timestamp of a commit and then
have ‘date’ reinterpret that datum with a given format string. This has two
issues:

- Reinterpreting the value with ‘date’ without specifying a time-zone creates
  possible inconsistencies in output due to developer time-zone settings.
- The -u option to ‘date’ is not actually portable.

Both of these issues can be avoided by just letting git to the whole job.

See the discussion in https://github.com/ft/ufw/pull/3 for details.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>
Reported-by: Jan Krautmacher <jan@krautmacher.org>