ep: Optional control for endpoint driver retry behaviour
The standard endpoint core handles EAGAIN, EINTR and empty transactions. User
may want to inject sideeffects into this behaviour, such as backing off in time
with regard to when retries happen or how different error codes are handled.

This allows for fine grained control of this behaviour. The default behaviour
of endpoints is not changed. It is further possible to initialise an optional,
arbitrary datum, that can keep track of the retry state of each individual
transaction.

The return codes that are controllable can be selected with via a bitmask. Un-
selected codes will exhibit their usual default behaviour.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>