Allow ufw to optionally be a Zephyr module
With these changes, all you really need to do is to include ufw in the
ZEPHYR_MODULES CMake variable. The setup_ufw() function supports this
transparently.

This will require a major version bump, because if you still want to use ufw in
a zephyr build link this, but without it being a zephyr module, you will need
to call setup_ufw() with the new FORCE_INCLUDE optional argument.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>