sx: Fix order of arguments to calloc()
New toolchains warn about this, and rightfully so. Standard calloc() wants the
item count as its first argument, and the item size as its second. I am not
sure how this snug in. There's no real code impact at the moment, because you
end up with the same amount of memory in either case. This order is the correct
one. From malloc(1):

…
SYNOPSIS
…
    void *calloc(size_t nmemb, size_t size);

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>