Optionally create a dynamic library with the full UFW API
This will be used for ABI/API compatibility checks. Normal users should only
use the static libraries with ufw. The corresponding option is therefore off by
default. The libufw-full library contains all of ufw, ufw-sx, and ufw-tap.

This changes CMAKE_SYSTEM_NAME in the gnu and clang toolchain files to Linux as
well, to enable building shared libraries — the Generic type does not allows
this.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>