ep: Fix warning with TI toolchains
The warning is this:

src/endpoints/core.c, line 540: warning: transfer of control bypasses initialization of:
                                variable "rest" (declared at line 547)

...which sounds like a false positive to me. While it is true, that the
initialisation *is* bypassed, the variable in question is not used between the
goto and the label to which is jumped. None of the other toolchains warn about
this either.

The changes are harmless, so we just do this to be warning free for the
upcoming release.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>