clang-tidy: Enable readability-named-parameter
I changed my stance on this a little. Techically I still think it's better to
only have names with the implementation, so you don't have to change multiple
places when you change names. But a lot of tooling gets confused by doing this,
sadly. So now it is "parameter names in header files" as well.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>