clang-tidy: Enable readability-uppercase-literal-suffix
Previously, ufw used lower-case literal suffixes. The readability rationale is,
that you can see the upper case ones better. It is arguable that this depends
on the situation really. For consistency, one could select one style and use
that throughout. Uppercase might be more useful more often, and that seems what
most code-styles have decided upon. Which is why this clang-tidy test exists.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>